% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{circadian-analysis}
\alias{circadian-analysis}
\title{Circadian Rhythm Analysis}
\description{
Non-parametric methods for characterizing 24-hour activity patterns.
}
\references{
\insertRef{witting1990}{actiRhythm}

\insertRef{vansomeren1999}{actiRhythm}

\insertRef{phillips2017}{actiRhythm}

\insertRef{wittmann2006}{actiRhythm}
}
