% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{circadian_cpp}
\alias{circadian_cpp}
\title{Complete Circadian Analysis}
\usage{
circadian_cpp(minute_data, hours_per_day = 24L, start_minute = 0L)
}
\arguments{
\item{minute_data}{Numeric vector of minute-level activity}

\item{hours_per_day}{Hours per day (default: 24)}

\item{start_minute}{Start minute of day (0-1439)}
}
\value{
List with L5, M10, RA, IS, IV, phi
}
\description{
Complete Circadian Analysis
}
