% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{.calculate.RA}
\alias{.calculate.RA}
\title{Calculate Relative Amplitude}
\usage{
.calculate.RA(M10, L5)
}
\arguments{
\item{M10}{Most active 10-hour average}

\item{L5}{Least active 5-hour average}
}
\value{
Numeric. Relative amplitude (0-1)
}
\description{
RA = (M10 - L5) / (M10 + L5)
}
\references{
van Someren EJ, et al. (1999). Chronobiol Int, 16(4):505-518.
}
\keyword{internal}
