% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{.classify.rhythm.pattern}
\alias{.classify.rhythm.pattern}
\title{Helper: Classify rhythm pattern based on harmonic contributions}
\usage{
.classify.rhythm.pattern(components)
}
\description{
Helper: Classify rhythm pattern based on harmonic contributions
}
\keyword{internal}
