% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{.empty.extended.cosinor.result}
\alias{.empty.extended.cosinor.result}
\title{Helper: Empty extended cosinor result}
\usage{
.empty.extended.cosinor.result(harmonics)
}
\description{
Helper: Empty extended cosinor result
}
\keyword{internal}
