% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{has.inclinometer}
\alias{has.inclinometer}
\title{Check if AGD Data Has Inclinometer Information}
\usage{
has.inclinometer(agd_data)
}
\arguments{
\item{agd_data}{List returned from read.agd()}
}
\value{
Logical. TRUE if inclinometer data is available
}
\description{
Checks whether an AGD file contains inclinometer data, which can be used
to enhance sleep detection algorithms.
}
