% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{IS_cpp}
\alias{IS_cpp}
\title{Interdaily Stability (IS)}
\usage{
IS_cpp(hourly_data, hours_per_day = 24L)
}
\arguments{
\item{hourly_data}{Numeric vector of hourly activity}

\item{hours_per_day}{Hours per day (default: 24)}
}
\value{
IS value (0-1, higher = more stable)
}
\description{
Interdaily Stability (IS)
}
