% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{IV_cpp}
\alias{IV_cpp}
\title{Intradaily Variability (IV)}
\usage{
IV_cpp(hourly_data)
}
\arguments{
\item{hourly_data}{Numeric vector of hourly activity}
}
\value{
IV value (lower = less fragmented)
}
\description{
Intradaily Variability (IV)
}
