% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{L5M10_cpp}
\alias{L5M10_cpp}
\title{Calculate L5/M10 Circadian Metrics}
\usage{
L5M10_cpp(minute_data, window_L5 = 300L, window_M10 = 600L, start_minute = 0L)
}
\arguments{
\item{minute_data}{Numeric vector of minute-level activity}

\item{window_L5}{L5 window in minutes (default: 300)}

\item{window_M10}{M10 window in minutes (default: 600)}

\item{start_minute}{Start minute of day (0-1439)}
}
\value{
List with L5_value, L5_onset, L5_onset_hours, M10_value, M10_onset,
M10_onset_hours, RA
}
\description{
Finds least active 5 hours (L5) and most active 10 hours (M10)
using van Someren (1999) average-profile method.
}
