% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{plot.actiRhythm_circadian}
\alias{plot.actiRhythm_circadian}
\title{Plot Circadian Rhythm Profile}
\usage{
\method{plot}{actiRhythm_circadian}(x, type = "profile", ...)
}
\arguments{
\item{x}{actiRhythm_circadian object from circadian.rhythm()}

\item{type}{Type of plot: "profile" (default), "daily", or "all"}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
ggplot object (or list of ggplot objects if type="all")
}
\description{
Plots the circadian rhythm analysis.
}
