% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{print.actiRhythm_circadian}
\alias{print.actiRhythm_circadian}
\title{Print Method for Circadian Rhythm Results}
\usage{
\method{print}{actiRhythm_circadian}(x, ...)
}
\arguments{
\item{x}{Object of class 'actiRhythm_circadian'}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Circadian Rhythm Results
}
