% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{print.actiRhythm_cosinor}
\alias{print.actiRhythm_cosinor}
\title{Print method for cosinor analysis}
\usage{
\method{print}{actiRhythm_cosinor}(x, ...)
}
\arguments{
\item{x}{actiRhythm_cosinor object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for cosinor analysis
}
