% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_fractal.R
\name{print.actiRhythm_dfa}
\alias{print.actiRhythm_dfa}
\title{Print Method for DFA Results}
\usage{
\method{print}{actiRhythm_dfa}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"actiRhythm_dfa"}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for DFA Results
}
