% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Rolling Mean}
\usage{
rolling_mean(x, window)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Window size}
}
\value{
Rolling means
}
\description{
Rolling Mean
}
