% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{rolling_sd}
\alias{rolling_sd}
\title{Rolling Standard Deviation}
\usage{
rolling_sd(x, window)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Window size}
}
\value{
Rolling SDs
}
\description{
Rolling Standard Deviation
}
