% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_actiRhythm}
\alias{scale_color_actiRhythm}
\title{ggplot2 Scale for actiRhythm Colors}
\usage{
scale_color_actiRhythm(type = "categorical", ...)
}
\arguments{
\item{type}{Character. Palette type (see actiRhythm_colors)}

\item{...}{Additional arguments passed to ggplot2::scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
Discrete color scale using the actiRhythm palette.
}
