% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_actiRhythm}
\alias{scale_fill_actiRhythm}
\title{ggplot2 Fill Scale for actiRhythm Colors}
\usage{
scale_fill_actiRhythm(type = "categorical", ...)
}
\arguments{
\item{type}{Character. Palette type (see actiRhythm_colors)}

\item{...}{Additional arguments passed to ggplot2::scale_fill_manual}
}
\value{
A ggplot2 scale object
}
\description{
Discrete fill scale using the actiRhythm palette.
}
