% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_actiRhythm_theme}
\alias{set_actiRhythm_theme}
\title{Set actiRhythm Theme as Default}
\usage{
set_actiRhythm_theme(...)
}
\arguments{
\item{...}{Arguments passed to theme_actiRhythm()}
}
\value{
\code{NULL}, invisibly; called for its side effect of setting
\code{theme_actiRhythm()} as the session default ggplot2 theme.
}
\description{
Sets theme_actiRhythm() as the default ggplot2 theme for the session.
}
