% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiRhythm-package.R
\docType{package}
\name{actiRhythm-package}
\alias{actiRhythm}
\alias{actiRhythm-package}
\title{actiRhythm: Circadian Rest-Activity Rhythm Analysis}
\description{
Quantifies the circadian rest-activity rhythm of a single subject from activity
counts: nonparametric measures, cosinor models with a rhythmicity test, period
estimation, fractal and nonlinear measures, the Sleep Regularity Index, and
phase metrics. Analyses run on an activity-count vector and its timestamps, and
a built-in reader loads 'ActiGraph' '.agd' files directly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdazadda/actiRhythm}
  \item \url{https://rdazadda.github.io/actiRhythm/}
  \item Report bugs at \url{https://github.com/rdazadda/actiRhythm/issues}
}

}
\author{
\strong{Maintainer}: Raymond Dacosta Azadda \email{rdazadda@alaska.edu}

Authors:
\itemize{
  \item Raymond Dacosta Azadda \email{rdazadda@alaska.edu}
  \item Andrew Grogan-Kaylor \email{agrogan@umich.edu}
  \item KyungSook Lee \email{klee98@alaska.edu}
}

}
\keyword{internal}
